Extracted middleware implementation
Extracted the middleware implementation from the file strategy.
To be able reuse it in other strategies
