+++ tests/unit/phpDocumentor/Reflection/Php/Factory/FileTest.php
@@ -15,8 +15,8 @@ namespace phpDocumentor\Reflection\Php\Factory;
 
 use Mockery as m;
 use phpDocumentor\Reflection\Fqsen;
+use phpDocumentor\Reflection\Middleware\Middleware;
 use phpDocumentor\Reflection\Php\Factory\File\LocalAdapter;
-use phpDocumentor\Reflection\Php\Factory\File\Middleware;
 use phpDocumentor\Reflection\Php\NodesFactory;
 use phpDocumentor\Reflection\Php\StrategyContainer;
 use phpDocumentor\Reflection\Php\File as FileElement;
